requirements.txt
streamlit
pandas
matplotlib
seaborn
scipy
langchain
langchain-experimental
langchain-community
langchain-groq
python-dotenv
wikipedia
tabulate
